==> 11a-master-sampling.ipynb <==
# Check the file path of sales data and import packages

## Check the file path

%pwd

%cd /data/p_dsi/teams2023/bridgestone_data/data/

%ls

## Import packages

import pandas as pd
import numpy as np
#import missingno as msno
import seaborn as sns
import matplotlib.pyplot as plt
import glob

# 1. Load product data and extract ARTICLE_ID for tire sales

## 1.1 Load product data

product = pd.read_csv('/data/p_dsi/teams2023/bridgestone_data/data/product.csv', sep = '|')

product

product.info()

## 1.2 Select tire sales data

tire_sales = product[(product["CATEGORY_CODE"] == 44.0) | (product["CATEGORY_CODE"] == 26.0)]
tire_sales.info()

## 1.3 Extract Article_ID

# Extract the ARTICLE_ID from tire_sales
ARTICLE_ID_list = tire_sales["ARTICLE_ID"].unique().tolist()

#print(ARTICLE_ID_list)
len(ARTICLE_ID_list)

# 2. Load sales data and subset all tire sales

## 2.1 Load sales data

# Get disired files
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2015*.csv')

nrow_tire_sale = []
nrow_total_sale = []
read_list = []
# Loop through the list of file names and read in each file
master_tire_sales_2015 = pd.DataFrame()
for file in csv_files:
    sales_2015 = pd.read_csv(file, sep = "|")
    read_list.append(file)
    nrow_total_sale.append(len(sales_2015))
    tire_sales_2015 = sales_2015[sales_2015['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    nrow_tire_sale.append(len(tire_sales_2015))
    master_tire_sales_2015 = pd.concat([master_tire_sales_2015, tire_sales_2015])
print(nrow_tire_sale)
print(nrow_total_sale)
print(read_list)

master_tire_sales_2015.to_csv("/data/p_dsi/teams2023/team7/tire_sales_2015.csv")

| tire sale | total sale |   year   |
|---------:|----------:|:--------:|
|   673029 |  16030135 | 2015-07  |
|   667178 |  16505764 | 2015-05  |
|   810755 |  15394337 | 2015-11  |
|   618789 |  14673428 | 2015-04  |
|   689657 |  15407864 | 2015-09  |
|   678131 |  16475323 | 2015-08  |
|   699409 |  15677601 | 2015-10  |
|   645751 |  14078243 | 2015-12  |
|   675639 |  16486393 | 2015-06  |


# Get disired files
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2016*.csv')

nrow_tire_sale = []
nrow_total_sale = []
read_list = []
# Loop through the list of file names and read in each file
master_tire_sales_2016 = pd.DataFrame()
for file in csv_files:
    sales_2016 = pd.read_csv(file, sep = "|")
    read_list.append(file)
    nrow_total_sale.append(len(sales_2016))
    tire_sales_2016 = sales_2016[sales_2016['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    nrow_tire_sale.append(len(tire_sales_2016))
    master_tire_sales_2016 = pd.concat([master_tire_sales_2016, tire_sales_2016])
print(nrow_tire_sale)
print(nrow_total_sale)
print(read_list)

master_tire_sales_2016.to_csv("/data/p_dsi/teams2023/team7/tire_sales_2016.csv")

| tire sale | total sale | year    |
|----------:|----------:|:--------|
|    629539 |  15074908 | 2016_01 |
|    671428 |  15480853 | 2016_03 |
|    727840 |  15153757 | 2016_09 |
|    647754 |  15008685 | 2016_04 |
|    693659 |  15537579 | 2016_08 |
|    688339 |  15462512 | 2016_10 |
|    726620 |  15650857 | 2016_06 |
|    745151 |  15019941 | 2016_12 |
|    676356 |  15524013 | 2016_07 |
|    787057 |  14882502 | 2016_11 |
|    667905 |  15480584 | 2016_05 |
|    617173 |  14737569 | 2016_02 |

# Get disired files
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2017*.csv')

nrow_tire_sale = []
nrow_total_sale = []
read_list = []
# Loop through the list of file names and read in each file
master_tire_sales_2017 = pd.DataFrame()
for file in csv_files:
    sales_2017 = pd.read_csv(file, sep = "|")
    read_list.append(file)
    nrow_total_sale.append(len(sales_2017))
    tire_sales_2017 = sales_2017[sales_2017['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    nrow_tire_sale.append(len(tire_sales_2017))
    master_tire_sales_2017 = pd.concat([master_tire_sales_2017, tire_sales_2017])
print(nrow_tire_sale)
print(nrow_total_sale)
print(read_list)

master_tire_sales_2017.to_csv("/data/p_dsi/teams2023/team7/tire_sales_2017.csv")

| tire sale | total sale | year    |
|----------:|----------:|:--------|
|    711740 |  15602358 | 2017_08 |
|    651629 |  15388366 | 2017_07 |
|    658973 |  15760198 | 2017_05 |
|    629243 |  14041986 | 2017_02 |
|    618192 |  14663301 | 2017_01 |
|    714905 |  15881014 | 2017_03 |
|    690634 |  15800806 | 2017_10 |
|    730871 |  15885612 | 2017_06 |
|    689522 |  14876691 | 2017_09 |
|    655209 |  14542920 | 2017_12 |
|    630170 |  14912797 | 2017_04 |
|    787040 |  15591791 | 2017_11 |

# Get disired files
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2018*.csv')

nrow_tire_sale = []
nrow_total_sale = []
read_list = []
# Loop through the list of file names and read in each file
master_tire_sales_2018 = pd.DataFrame()
for file in csv_files:
    sales_2018 = pd.read_csv(file, sep = "|")
    read_list.append(file)
    nrow_total_sale.append(len(sales_2018))
    tire_sales_2018 = sales_2018[sales_2018['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    nrow_tire_sale.append(len(tire_sales_2018))
    master_tire_sales_2018 = pd.concat([master_tire_sales_2018, tire_sales_2018])
print(nrow_tire_sale)
print(nrow_total_sale)
print(read_list)

master_tire_sales_2018.to_csv("/data/p_dsi/teams2023/team7/tire_sales_2018.csv")

| tire sale | total sale |   year  |
|---------:|----------:|:-------:|
|   645927 |  15197671 | 2018_07 |
|   670846 |  15630241 | 2018_08 |
|   656858 |  15942240 | 2018_05 |
|   606078 |  14449096 | 2018_04 |
|   722685 |  15888468 | 2018_06 |
|   611146 |  13247533 | 2018_02 |
|   600992 |  14986650 | 2018_01 |
|   682535 |  15432276 | 2018_10 |
|   718014 |  14804703 | 2018_09 |
|   734588 |  16057725 | 2018_03 |

# 3. Resample each sale.csv and combine them into a master dataframe

### *Note*
### The total number of tire sales from 2015-2018 is 29254956, and we want to have 100000 samples in our master sample data. The ratio calculation is 100000/29254956 = 0.00341822424. 

master_df = pd.DataFrame()
file_names = glob.glob('/data/p_dsi/teams2023/team7/*tire_sales*.csv')

for file_name in file_names:
    file = pd.read_csv(file_name, sep = ",")
    file_sampled = file.sample(frac=0.00341822424, replace=False, random_state=1)
    master_df = pd.concat([master_df, file_sampled])

master_df.to_csv("/data/p_dsi/teams2023/team7/sales_sample.csv")

# 4. Join sales_sample.csv with vehicle.csv, store.csv, individual.csv and write master_sample.csv

## 4.1 Read in data

tire_sales_sampled = pd.read_csv("/data/p_dsi/teams2023/team7/sales_sample.csv")

individual = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/individual.csv")
# product read before
store = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/store.csv", sep = "|")
vehicle = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/vehicle.csv", sep = "|")

## 4.2 Merge

final_df = tire_sales_sampled.merge(individual, left_on = 'INDIV_ID', right_on = "MZB_INDIV_ID", how = 'left')

final_df = final_df.merge(product, on = "ARTICLE_ID", how = 'left')

final_df = final_df.merge(store, on = "STORE_ID", how = 'left')

final_df = final_df.merge(vehicle, on = "VEHICLE_ID", how = 'left')

## 4.3 Sanity check

pd.set_option('display.max_columns', None)

final_df.sample(10)

print("tire_sales_sampled:\n", tire_sales_sampled.columns, 
      "\n", "\nindividual:\n", individual.columns, 
      "\n", "\nproduct:\n", product.columns, 
      "\n", "\nstore:\n", store.columns, 
      "\n", "\nvehicle:\n", vehicle.columns)

print("\nTotal number of columns:", len(tire_sales_sampled.columns) 
      + len(individual.columns)
      + len(product.columns)
      + len(store.columns)
      + len(vehicle.columns))

print("final_df:\n", final_df.columns)
print("\nTotal number of columns:", len(final_df.columns))

## 4.4 Write out

final_df.to_csv("/data/p_dsi/teams2023/team7/master_sample.csv")
