==> 11b-master-sampling-part2.ipynb <==
# 1. Import packages

import pandas as pd
import numpy as np
import seaborn as sns
import glob

# 2. Load product data and extract ARTICLE_ID for tire sales

## 2.1 Load product data

product = pd.read_csv('/data/p_dsi/teams2023/bridgestone_data/data/product.csv', sep = '|')

product.info()
product.sample(10)

product.info()

## 2.2 Select tire sales data

tire_sales = product[(product["CATEGORY_CODE"] == 44.0) | (product["CATEGORY_CODE"] == 26.0)]
tire_sales.info()

## 2.3 Extract Article_ID

# Extract the ARTICLE_ID from tire_sales
ARTICLE_ID_list = tire_sales["ARTICLE_ID"].unique().tolist()

#print(ARTICLE_ID_list)
len(ARTICLE_ID_list)

# 3. Load sales data and subset all tire sales

# sample testing with one month in 2017
sales_2017_oct = pd.read_csv('/data/p_dsi/teams2023/bridgestone_data/data/sales_20171031.csv', sep = '|')
sales_2017_oct.info()

# example of getting ARTICLE_ID and UNITS 
master_small = sales_2017_oct[["ARTICLE_ID", "UNITS"]]
tire_sales_2017 = master_small[master_small['ARTICLE_ID'].isin(ARTICLE_ID_list)]
tire_sales_2017
df = tire_sales_2017.groupby(by=["ARTICLE_ID"]).sum()
df.info()

# Create a concatenated data frame with ARTICLE_ID and UNITS in Nov, Dec 2017
csv_files = ['/data/p_dsi/teams2023/bridgestone_data/data/sales_20171130.csv',
             '/data/p_dsi/teams2023/bridgestone_data/data/sales_20171231.csv']

# create a new data frame
master_tire_sales_2017 = pd.DataFrame()
for file in csv_files:
    sales_2017 = pd.read_csv(file, sep = "|")
    sales_small = sales_2017[["ARTICLE_ID", "UNITS"]]
    tire_sales_2017 = sales_small[sales_small['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    df = tire_sales_2017.groupby(by=["ARTICLE_ID"]).sum()
    master_tire_sales_2017 = pd.concat([master_tire_sales_2017, df])

master_tire_sales_2017.info()
master_tire_sales_2017.sample(10)

# Sum by the article_id again to get the total result
master_tire_sales_2017 = master_tire_sales_2017.groupby(by=["ARTICLE_ID"]).sum()
master_tire_sales_2017.info()
master_tire_sales_2017.sample(10)

# save as csv file
master_tire_sales_2017.to_csv("/data/p_dsi/teams2023/team7/tire_sales_article_id_and_total_units_Nov_Dec_2017.csv")

# Create a concatenated data frame with ARTICLE_ID and UNITS from Jan - Oct in 2018
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2018*.csv')

master_tire_sales_2018 = pd.DataFrame()
for file in csv_files:
    sales_2018 = pd.read_csv(file, sep = "|")
    sales_small = sales_2018[["ARTICLE_ID", "UNITS"]]
    tire_sales_2018 = sales_small[sales_small['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    df = tire_sales_2018.groupby(by=["ARTICLE_ID"]).sum()
    master_tire_sales_2018 = pd.concat([master_tire_sales_2018, df])
    
master_tire_sales_2018.info()
master_tire_sales_2018.sample(10)

# Sum by the article_id again to get the total result
master_tire_sales_2018 = master_tire_sales_2018.groupby(by=["ARTICLE_ID"]).sum()
master_tire_sales_2018.info()
master_tire_sales_2018.sample(10)

# save as csv file
master_tire_sales_2018.to_csv("/data/p_dsi/teams2023/team7/tire_sales_article_id_and_total_units_2018.csv")

# Now load two files from 2017 and 2018 and bind them together for the last time
unit_tire_sales_year = pd.concat([master_tire_sales_2017, master_tire_sales_2018])
unit_tire_sales_year = unit_tire_sales_year.groupby(by=["ARTICLE_ID"]).sum()

# filter by the sales greater than 5,000
unit_tire_sales_year = unit_tire_sales_year.loc[unit_tire_sales_year['UNITS'] > 5000]
unit_tire_sales_year.info()
unit_tire_sales_year.sample(10)

#save the final result in csv file
unit_tire_sales_year.to_csv("/data/p_dsi/teams2023/team7/tire_sales_article_id_and_total_units_last_12_months.csv")

# Sanity check
unit_tire_sales_last_12_months = pd.read_csv("/data/p_dsi/teams2023/team7/tire_sales_article_id_and_total_units_last_12_months.csv", sep = ',')

# Extract the ARTICLE_ID from tire_sales
ARTICLE_ID_list = unit_tire_sales_last_12_months["ARTICLE_ID"].tolist()

#print(ARTICLE_ID_list)
len(ARTICLE_ID_list)

master_sample = pd.read_csv("/data/p_dsi/teams2023/team7/master_sample.csv")

master_sample.info()
master_sample.sample(10)

master_sample_selected_article_ids = master_sample[master_sample['ARTICLE_ID'].isin(ARTICLE_ID_list)]

master_sample_selected_article_ids.info()
master_sample_selected_article_ids.sample(10)
