==> 12a-data-validation.ipynb <==
import pandas as pd
import glob

# Data Validation for sales_2015

csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2015*.csv')

df = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/sales_20150430.csv", sep = "|")

ref_columns = set(df.columns)
print(ref_columns)

for df in csv_files:
    df = pd.read_csv(df, sep = "|")
    if not set(df.columns) == ref_columns:
        print(f"Data frame {df} does not have the same columns as the reference set.")
    else:
        print("Columns are identical!")

# Data Validation for sales_2016

# load the files of sales data for the entire 2016
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2016*.csv')

# read in one csv file that will be used to create a set of column names
df = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/sales_20160131.csv", sep = "|")

# create a reference column
ref_columns = set(df.columns)
print(ref_columns)

for df in csv_files:
    df = pd.read_csv(df, sep = "|")
    if not set(df.columns) == ref_columns:
        print(f"Data frame {df} does not have the same columns as the reference set.")
    else:
        print("Columns are identical!")

# Data Validation for sales_2017

# load the files of sales data for the entire 2017
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2017*.csv')

# read in one csv file that will be used to create a set of column names
df = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/sales_20170131.csv", sep = "|")

# create a reference column
ref_columns = set(df.columns)
print(ref_columns)

for df in csv_files:
    df = pd.read_csv(df, sep = "|")
    if not set(df.columns) == ref_columns:
        print(f"Data frame {df} does not have the same columns as the reference set.")
    else:
        print("Columns are identical!")

# Data Validation for sales_2018

# load the files of sales data for the entire 2017
csv_files = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_2018*.csv')

# read in one csv file that will be used to create a set of column names
df = pd.read_csv("/data/p_dsi/teams2023/bridgestone_data/data/sales_20180131.csv", sep = "|")
