==> 13-monthly-tire-sale.ipynb <==
# 1. Import Packages
----------

import pandas as pd
import glob
import os

%cd /data/p_dsi/teams2023/bridgestone_data/data/

%ls

# 2. Extract all ARTICLE_IDs that are relative to tire sales
----------------------

product = pd.read_csv('product.csv', sep = '|')

tire_sales = product[(product["CATEGORY_CODE"] == 44.0) | (product["CATEGORY_CODE"] == 26.0)]

ARTICLE_ID_list = tire_sales["ARTICLE_ID"].unique().tolist()

# 3. Get the path for every monthly sale.csv
--------

paths = glob.glob('/data/p_dsi/teams2023/bridgestone_data/data/*sales_201*.csv')

csv_files = []

#for path in paths:
for file in os.listdir("/data/p_dsi/teams2023/bridgestone_data/data"):
    if file.startswith('sales_201'):
        csv_files.append(file)

print(csv_files)
print(len(csv_files))

# 4. Write out monthly tire sales
-----------

for file_name in csv_files:
    in_path = "/data/p_dsi/teams2023/bridgestone_data/data/" + file_name
    sales = pd.read_csv(in_path, sep = "|")
    tire_sales = sales[sales['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    out_path = "/data/p_dsi/teams2023/team7/tire_" + file_name
    print(out_path)
    tire_sales.to_csv(out_path)
