==> 14-data-cleaning-drop-duplicates.ipynb <==
# Drop duplicates from sales data

The raw sales data contains an estimated 10% of duplicate values. This notebook will process the data to remove them, and create new datasets for future use in EDA and modeling. Runtimes are also taken and printed to test and track efficiency. 

The output from this notebook are 43 csv files, which are stored in the team7/removed-duplicates directory.

## Install packages

import glob
import numpy as np
import os
import pandas as pd

## Load data

# column names only
header_only = pd.read_csv("sales_20180228.csv", sep="|", nrows=0)
cols = header_only.columns.to_list()
cols

# nrows = 1M for testing
%cd /data/p_dsi/teams2023/bridgestone_data/data

dat_test_tiny = pd.read_csv("sales_20180228.csv", sep="|", nrows=1000000, parse_dates=["DATE"])
dat_test_tiny.shape

%%time
%cd /data/p_dsi/teams2023/bridgestone_data/data

dat_test_full = pd.read_csv("sales_20180228.csv", sep="|", parse_dates=["DATE"])
dat_test_full.shape

%%time
dat_test_full = pd.read_csv("sales_20180228.csv", sep="|", parse_dates=["DATE"])

## Custom Functions


def duplicates_test(data_raw, data_dropped):
    '''
    Check whether duplicates are properly removed using drop_duplicates out-of-the-box on the raw sales data.
    
    data_raw = original sales dataframe
    data_dropped = dataframe of month of sales data after dropping duplicates
    '''
    # group data and add count column
    cols = data_raw.columns.tolist()
    temp = data_raw.groupby(cols).size().reset_index(name="count")
    target_nrows = temp.shape[0]   # target = nrows after grouping
    
    # run test:
    print(". . . \n")
    print("Results from duplicates_test \n")
    if data_dropped.shape[0] == target_nrows: print("PASS")
    else: print("FAIL: nrows do not match")
    
    # sanity check - print descriptive info
    print("Original nrows: ", data_raw.shape[0])
    print("Target nrows: ", target_nrows) 
    print(". . . \n")
    print("Nrows after drop_duplicates: ", data_dropped.shape[0])
    
    return

def drop_duplicates_big_data(bigdata, cols, filename): 
    '''
    Given a big data dataframe, removes duplicates. Returns 'new_df'.
    
    bigdata = pandas dataframe of the big dataset
    cols = list of sales columns
    filename = passed in from looping over csv_list
    '''
    temp = bigdata.groupby(cols).size().reset_index(name="count")
    temp_nonduplicate_obs = temp.query("count == 1").drop(["count"], axis=1)
    temp_duplicate_obs = temp.query("count > 1").drop(["count"], axis=1)
    
    # df with duplicates removed
    temp_drop_duplicates = temp_duplicate_obs.drop_duplicates()
    # concat dfs with distinct obs
    new_df = pd.concat([temp_nonduplicate_obs, temp_drop_duplicates], axis=0)

# print message for testing only:    
#     print(". . . \n")
#     print("Results from drop_duplicates_big_data \n")
#     print("Shape of new df (after dropping duplicates): ", new_df.shape)
#     print(". . . \n")
    
    return new_df 

### Test drop_duplicates_big_data on 1M rows of data

%%time
drop_test = drop_duplicates_big_data(dat_test_tiny)

duplicates_test(data_raw=dat_test_tiny, data_dropped=drop_test)

## Test drop_duplicates_big_data on one sales file

Test on pre-loaded dataframe

%%time
drop_test = drop_duplicates_big_data(dat_test_full)

duplicates_test(data_raw=dat_test_full, data_dropped=drop_test)

Now write for loop to operate across all 43 sales csv files 

%%time
cols = drop_test.columns.to_list()
drop_test.sort_values(cols).groupby(cols).size().reset_index(name="count").query("count == 1").shape[0]

%cd /data/p_dsi/teams2023/team7

csv_files = []
for file in os.listdir("/data/p_dsi/teams2023/bridgestone_data/data"):
    if file.startswith('sales_2'):
        csv_files.append(file)

if len(csv_files) != 43: print("ERROR - csv_files length = ", len(csv_files))

%%time
# store col names to be accessed in the loop
cols = dat_test_tiny.columns.to_list()
# loop counter variable
count = 1

for filename in csv_files:
    # store in and out file paths as variables
    path_in = "/data/p_dsi/teams2023/bridgestone_data/data/" + filename
    path_out = "/data/p_dsi/teams2023/team7/remove-duplicates/" + filename
    
    # read in one sales csv as panda dataframe
    df_read = pd.read_csv(path_in, sep = "|", parse_dates=["DATE"])
    
    # drop duplicates and store in new df
    temp = drop_duplicates_big_data(bigdata=df_read, cols=cols)
    
    # write new df to csv
    temp.to_csv(path_out, index=False)
    
    # status message with loop count
    print("fParsing complete: #{count}, {filename}")
    count += 1
    
    break

from pyarrow import csv

parse_opts = csv.ParseOptions(delimiter="|")

%%time
test_pa = csv.read_csv("sales_20180228.csv", parse_options=parse_opts)

%%time
test_convert_pd = test_pa.to_pandas()

Test run the for loop...

%%time
# set options to read csv w/pyarrow
parse_opts = csv.ParseOptions(delimiter="|")
read_opts = csv.ReadOptions(use_threads=4)
# loop counter variable
count = 1
# dict to store nrows of each iteration
results_nrows = {}

for filename in csv_files:
    # store in and out file paths as variables
    path_in = "/data/p_dsi/teams2023/bridgestone_data/data/" + filename
    path_out = "/data/p_dsi/teams2023/team7/remove-duplicates/" + filename
    
    # read in one sales csv as Arrow table
    arrow_table = csv.read_csv(path_in, parse_options=parse_opts)
    # convert Arrow table to pandas df
    df = arrow_table.to_pandas()
    
    # drop duplicates and store in new df
    temp = drop_duplicates_big_data(bigdata=df, cols=cols, filename=filename)
    
    # store the nrows from each df in a dict
    results_nrows.update({f"{filename}": temp.shape[0]})
    
    # write new df to csv
    temp.to_csv(path_out, index=False)
    
    # status message with loop count
    print(f"Parsing complete: #{count}, {filename}")
    count += 1
    print(results_nrows)
    break

%%time
# sanity check on above loop test
duplicates_test(data_raw=df, data_dropped=temp)

Run full loop

%%time
# set options to read csv w/pyarrow
parse_opts = csv.ParseOptions(delimiter="|")
read_opts = csv.ReadOptions(use_threads=4)
# loop counter variable
count = 1
# dict to store nrows of each iteration
results_nrows = {}

for filename in csv_files:
    # store in and out file paths as variables
    path_in = "/data/p_dsi/teams2023/bridgestone_data/data/" + filename
    path_out = "/data/p_dsi/teams2023/team7/remove-duplicates/" + filename
    
    # read in one sales csv as Arrow table
    arrow_table = csv.read_csv(path_in, parse_options=parse_opts)
    # convert Arrow table to pandas df
    df = arrow_table.to_pandas()
    
    # drop duplicates and store in new df
    temp = drop_duplicates_big_data(bigdata=df, cols=cols, filename=filename)
    
    # write new df to csv
    temp.to_csv(path_out, index=False)
    
    # status message with loop count
    print(f"Processed: File no. {count}, File name: {filename}")
    count += 1
print("Parsing complete, check directory for new csvs.")

#### Check one of the new files...

df = pd.read_csv("/data/p_dsi/teams2023/team7/remove-duplicates/sales_20181031.csv", nrows=10)
df
