==> 16-create-sales-csvs-relevant-to-us.ipynb <==
### Imports

import pandas as pd
import os
import time

### Get the article_id list that we need

article_id_units = pd.read_csv("/data/p_dsi/teams2023/team7/tire_sales_article_id_and_total_units_last_12_months.csv", sep = ',')
article_ids = article_id_units["ARTICLE_ID"].tolist()

len(article_ids)

### Get all the sales file names

csv_files = []

for file in os.listdir("/data/p_dsi/teams2023/bridgestone_data/data"):
    if file.startswith('sales_201'):
        csv_files.append(file)

print(csv_files)
print(len(csv_files))

### Write final_sales_{}.csv for all months

start_time = time.time()
final_sales_dict = dict()

for file_name in csv_files:
    in_path = "/data/p_dsi/teams2023/bridgestone_data/data/" + file_name
    sales = pd.read_csv(in_path, sep = "|")
    final_sales = sales[sales['ARTICLE_ID'].isin(article_ids)]
    final_sales_dict[file_name] = final_sales
    out_path = "/data/p_dsi/teams2023/team7/final_" + file_name
    print(out_path)
    final_sales.to_csv(out_path)

end_time = time.time()
time_passed = end_time - start_time
print("time_passed:", time_passed)

### Write final_sales_2015.csv, final_sales_2016.csv, final_sales_2017.csv, and final_sales_2018.csv

years = [2015, 2016, 2017, 2018]
[pd.concat([final_sales_dict[key] for key in final_sales_dict.keys() if key.startswith("sales_{}".format(year))]).to_csv("/data/p_dsi/teams2023/team7/final_sales_{}.csv".format(year)) for year in years]

### Write final_sales.csv

pd.concat(final_sales_dict.values()).to_csv("/data/p_dsi/teams2023/team7/final_sales.csv")
