==> 17-final-sales-for-tires.ipynb <==
## 1. Import packages

import pandas as pd
import glob
import os
import time

%cd /data/p_dsi/teams2023/team7/remove-duplicates/

## 2. Subset tire sales

### 2.1 Subset by CATEGORY_CODE

product = pd.read_csv('/data/p_dsi/teams2023/bridgestone_data/data/product.csv', sep = '|')

tire_sales = product[(product["CATEGORY_CODE"] == 44.0) | (product["CATEGORY_CODE"] == 26.0)]

### 2.2 Get all the ariticle_ids for tire sales

ARTICLE_ID_list = tire_sales["ARTICLE_ID"].unique().tolist()

csv_files = ['sales_20171130.csv','sales_20171231.csv']

master_tire_sales_2017 = pd.DataFrame()
for file in csv_files:
    sales_2017 = pd.read_csv(file)
    sales_small = sales_2017[["ARTICLE_ID", "UNITS"]]
    tire_sales_2017 = sales_small[sales_small['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    df = tire_sales_2017.groupby(by=["ARTICLE_ID"]).sum()
    master_tire_sales_2017 = pd.concat([master_tire_sales_2017, df])

master_tire_sales_2017.info()

master_tire_sales_2017 = master_tire_sales_2017.groupby(by=["ARTICLE_ID"]).sum()
master_tire_sales_2017.info()

csv_files = glob.glob('*sales_2018*.csv')

master_tire_sales_2018 = pd.DataFrame()
for file in csv_files:
    sales_2018 = pd.read_csv(file)
    print(file)
    sales_small = sales_2018[["ARTICLE_ID", "UNITS"]]
    tire_sales_2018 = sales_small[sales_small['ARTICLE_ID'].isin(ARTICLE_ID_list)]
    df = tire_sales_2018.groupby(by=["ARTICLE_ID"]).sum()
    master_tire_sales_2018 = pd.concat([master_tire_sales_2018, df])

master_tire_sales_2018.info()

### 2.3 Get the article_ids whose sales are greater than 5000 from 2017 to 2018

unit_tire_sales_year = pd.concat([master_tire_sales_2017, master_tire_sales_2018])
unit_tire_sales_year = unit_tire_sales_year.groupby(by=["ARTICLE_ID"]).sum()

unit_tire_sales_year = unit_tire_sales_year.loc[unit_tire_sales_year['UNITS'] > 5000]
unit_tire_sales_year.info()

### 2.4 Save the aritcle_id list

#save the final result in csv file
unit_tire_sales_year.to_csv("/data/p_dsi/teams2023/team7/no_duplicate_tire_sales_article_id_and_total_units_last_12_months.csv")

article_id_units = pd.read_csv("/data/p_dsi/teams2023/team7/no_duplicate_tire_sales_article_id_and_total_units_last_12_months.csv", sep = ',')
article_ids = article_id_units["ARTICLE_ID"].tolist()

len(article_ids)

## 3. Save the final_sales CSVs for tire sales.

csv_files = []

for file in os.listdir("/data/p_dsi/teams2023/team7/remove-duplicates/"):
    if file.startswith('sales_201'):
        csv_files.append(file)

print(csv_files)
print(len(csv_files))

start_time = time.time()
no_duplicate_final_sales_dict = dict()

for file_name in csv_files:
    in_path = "/data/p_dsi/teams2023/team7/remove-duplicates/" + file_name
    sales = pd.read_csv(in_path)
    no_duplicate_final_sales = sales[sales['ARTICLE_ID'].isin(article_ids)]
    no_duplicate_final_sales_dict[file_name] = no_duplicate_final_sales
    out_path = "/data/p_dsi/teams2023/team7/no_duplicate_final_" + file_name
    print(out_path)
    no_duplicate_final_sales.to_csv(out_path)

end_time = time.time()
time_passed = end_time - start_time
print("time_passed:", time_passed)

years = [2015, 2016, 2017, 2018]
[pd.concat([no_duplicate_final_sales_dict[key] for key in no_duplicate_final_sales_dict.keys() if key.startswith("sales_{}".format(year))]).to_csv("/data/p_dsi/teams2023/team7/no_duplicate_final_sales_{}.csv".format(year)) for year in years]

pd.concat(no_duplicate_final_sales_dict.values()).to_csv("/data/p_dsi/teams2023/team7/no_duplicate_final_sales.csv")
